requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
pandas==2.2.3
openpyxl==3.1.5
django-import-export==4.1.0
requests==2.28.2
fastapi==0.95.1
uvicorn==0.21.0
torchvision==0.24.0
torch==2.9.0 --index-url https://download.pytorch.org/whl/cpu
tokenizers==0.15.2
transformers==4.38.2
